Add Activity stuff to Activity.py
